refactor: move 7 identity modules into identity/
