+++ tests/test_expand.py
@@ -17,7 +17,7 @@ import pytest
 from fastapi.testclient import TestClient
 
 from helix_context.retrieval import expand
-from helix_context import session_delivery
+from helix_context.identity import session_delivery
 from helix_context.config import (
     BudgetConfig,
     GenomeConfig,
